ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
opencv-python==4.8.1.78
librosa==0.10.1
pyaudioanalysis==0.3.14
deepface==0.0.79
tensorflow==2.15.0
